requirements.txt
streamlit
ollama
python-dotenv
langchain
fastapi
uvicorn
python-multipart
langchain-chroma
langchain-ollama
langchain-community
langchain-huggingface
transformers
torch 
accelerate
chromadb
torchaudio
faster-whisper 
pyannote.audio